pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scrubjay"
version = "0.1.0"
description = "Sanitize sensitive data in security logs before sending to LLM APIs"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "esinta, LLC", email = "oss@esinta.com" },
]
keywords = ["security", "sanitization", "llm", "splunk", "okta", "privacy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
mcp = []
claude = []
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.0",
]
all = [
    "scrubjay[mcp,claude,dev]",
]

[project.scripts]
scrubjay = "scrubjay.cli.main:cli"

[project.urls]
Homepage = "https://github.com/esinta/scrubjay"
Repository = "https://github.com/esinta/scrubjay"
Issues = "https://github.com/esinta/scrubjay/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/scrubjay"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
